Split lib.rs into several modules
