+++ src/embeddings_wrap.rs
@@ -1,0 +1,24 @@
+use finalfusion::prelude::*;
+
+pub enum EmbeddingsWrap {
+    NonView(Embeddings<VocabWrap, StorageWrap>),
+    View(Embeddings<VocabWrap, StorageViewWrap>),
+}
+
+impl EmbeddingsWrap {
+    pub fn storage(&self) -> &Storage {
+        use EmbeddingsWrap::*;
+        match self {
+            NonView(e) => e.storage(),
+            View(e) => e.storage(),
+        }
+    }
+
+    pub fn vocab(&self) -> &VocabWrap {
+        use EmbeddingsWrap::*;
+        match self {
+            NonView(e) => e.vocab(),
+            View(e) => e.vocab(),
+        }
+    }
+}
